+++ nuxeo/endpoint.py
@@ -72,7 +72,7 @@ class APIEndpoint(object):
                 return response.content
             json = response.json()
         else:
-            json = response
+            return response
 
         if cls is dict:
             return json
@@ -109,6 +109,8 @@ class APIEndpoint(object):
         response = self.client.request(
             'POST', endpoint, data=resource, raw=raw, **kwargs)
 
+        if isinstance(response, dict):
+            return response
         return self._cls.parse(response.json(), service=self)
 
     def put(self, resource=None, path=None, **kwargs):