+++ tests/manual/many_small_files.py
@@ -3,7 +3,6 @@ from __future__ import unicode_literals
 
 import os
 import random
-
 from memory_profiler import profile
 
 from nuxeo.client import Nuxeo