+++ tests/test_repository.py
@@ -2,7 +2,6 @@
 from __future__ import unicode_literals
 
 import operator
-
 import pytest
 
 from nuxeo.compat import get_bytes, get_error_message
@@ -141,7 +140,7 @@ def test_fetch_acls(server):
         assert acls[0]['name'] == 'inherited'
         aces = list(sorted(acls[0]['aces'], key=operator.itemgetter('id')))
         assert aces[0]['id'] == 'Administrator:Everything:true:::'
-        assert aces[1]['id'] == 'members:Read:true:::'
+        assert aces[-1]['id'] == 'members:Read:true:::'
 
 
 def test_fetch_blob(server):